Added routes to all screens, changed name of WordGridGameScreen to MatchingScreen
