+++ lib/core/router/routes.dart
@@ -1,5 +1,15 @@
 enum Routes {
-  initial('/');
+  initial('/'),
+  profile('/profile'),
+  allCourses('/all_courses'),
+  myCourses('/my_courses'),
+  course('/course'),
+  deepLearning('/deep_learning'),
+  test('/test'),
+  matching('/matching'),
+  dragAndDrop('/drag_and_drop'),
+  writing('/writing');
+
 
   final String path;
   const Routes(this.path);
